requirements.txt
# pip install -r requirements.txt
requests==2.24.0
biopython==1.78
beautifulsoup4==4.9.3
parasail==1.2
scikit-bio==0.5.6